+++ tests/autofix_files/async91x_py311.py
@@ -7,24 +7,20 @@ ASYNC913: indefinite-loop-no-guaranteed-checkpoint
 async912 handled in separate file
 """
 
-# ARG --enable=ASYNC910,ASYNC911,ASYNC913
+# ARG --enable=ASYNC910,ASYNC911,ASYNC913,ASYNC914
 # AUTOFIX
-# ASYNCIO_NO_AUTOFIX
 import trio
-
-
-async def foo(): ...
 
 
 async def foo_try_except_star_1():  # ASYNC910: 0, "exit", Statement("function definition", lineno)
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except* ValueError:
         ...
     except* RuntimeError:
         raise
     else:
-        await foo()
+        pass
     await trio.lowlevel.checkpoint()
 
 
@@ -34,12 +30,12 @@ async def foo_try_except_star_2():  # safe
     except* ValueError:
         ...
     finally:
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 async def foo_try_except_star_3():  # safe
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except* ValueError:
         raise
 
@@ -47,9 +43,9 @@ async def foo_try_except_star_3():  # safe
 # Multiple except* handlers - should all guarantee checkpoint/raise
 async def foo_try_except_star_4():
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except* ValueError:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except* TypeError:
         raise
     except* Exception:
@@ -62,7 +58,7 @@ async def try_else_no_raise_in_except():  # ASYNC910: 0, "exit", Statement("func
     except* ValueError:
         ...
     else:
-        await foo()
+        await trio.lowlevel.checkpoint()
     await trio.lowlevel.checkpoint()
 
 
@@ -72,12 +68,12 @@ async def try_else_raise_in_except():
     except* ValueError:
         raise
     else:
-        await foo()
+        await trio.lowlevel.checkpoint()
 
 
 async def check_async911():  # ASYNC911: 0, "exit", Statement("yield", lineno+7)
     try:
-        await foo()
+        await trio.lowlevel.checkpoint()
     except* ValueError:
         ...
     except* RuntimeError:
@@ -91,7 +87,118 @@ async def check_async913():
     while True:  # ASYNC913: 4
         await trio.lowlevel.checkpoint()
         try:
+            # If this was lowlevel it'd be marked as ASYNC914 after we insert a checkpoint
+            # before the try. TODO: Ideally it'd be fixed in the same pass.
             await foo()
         except* ValueError:
             # Missing checkpoint
             ...
+
+
+# ASYNC914
+async def foo_try_1():
+    await trio.lowlevel.checkpoint()
+    try:
+        pass
+    except* BaseException:
+        pass
+
+
+async def foo_try_2():
+    await trio.lowlevel.checkpoint()
+    try:
+        pass
+    except* BaseException:
+        await foo()
+
+
+async def foo_try_3():
+    await trio.lowlevel.checkpoint()
+    try:
+        await foo()
+    except* BaseException:
+        pass
+
+
+async def foo_try_4():
+    await trio.lowlevel.checkpoint()
+    try:
+        await foo()
+    except* BaseException:
+        await foo()
+
+
+async def foo_try_5():
+    await foo()
+    try:
+        pass
+    except* BaseException:
+        pass
+
+
+async def foo_try_6():
+    await foo()
+    try:
+        pass
+    except* BaseException:
+        await foo()
+
+
+async def foo_try_7():
+    await foo()
+    try:
+        await foo()
+    except* BaseException:
+        pass
+
+
+async def foo_try_8():
+    await foo()
+    try:
+        await foo()
+    except* BaseException:
+        await foo()
+
+
+async def foo_try_9():
+    try:
+        pass
+    except* BaseException:
+        await foo()
+    else:
+        await foo()
+
+
+async def foo_try_10():
+    try:
+        await trio.lowlevel.checkpoint()
+    finally:
+        await foo()
+
+
+async def foo_try_11():
+    try:
+        await trio.lowlevel.checkpoint()
+    except* BaseException:
+        await foo()
+
+
+async def foo_try_12():
+    try:
+        pass
+    except* BaseException:
+        ...
+    else:
+        await foo()
+    await trio.lowlevel.checkpoint()
+
+
+async def foo_try_13():
+    try:
+        ...
+    except* ValueError:
+        ...
+    except* BaseException:
+        raise
+    finally:
+        await trio.lowlevel.checkpoint()